refactor(asciicast): reorganize module structure with Transform trait
Extract types to dedicated types.rs with comprehensive rustdoc, add
Transform trait for future event transformations, remove unused
terminal preview code, and improve documentation throughout.

Changes:
- Create types.rs with Header, Event, EventType, AsciicastFile
- Create transform.rs with Transform trait and TransformChain
- Remove dead terminal_preview_at/styled_preview_at methods
- Add module-level docs and error documentation to reader/writer
- Reduce mod.rs to re-exports only (~30 lines)
